rename integrationtest folder Api to B2BApi
